requirements.txt
asgiref==3.8.1
boto3==1.34.0
botocore==1.34.162
Django==5.2.3
django-cors-headers==4.7.0
django-storages==1.14.2
jmespath==1.0.1
pillow==10.4.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
s3transfer==0.9.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
